Changed the token backing functionality and fixed a small mint bug
Changed the token backing functionality in a way that allows it to be extensible and allows adding new tokens in the future. As a byproduct, assets now cost less RAM when no assets are backed but cost more RAM when assets are backed. Also, backing assets needs to be announced using the "backsymbol" action

Also, now checks if the account of the new owner exists when minting an asset.